added baro alt & angle to UI over WiFi

+++ UI/Artificial Horizon/artificial_horizon/altimeter.pde
@@ -1,5 +1,5 @@
 class altimeter {
-  void display(int alt) {
+  void display(float alt) {
     fill(255);
     stroke(0);
     rect(500,100,80,400);
